Packages
Organised classes into suitable packages. Separated Main into new file.
