+++ src/main/java/Main.java
@@ -1,0 +1,11 @@
+import duke.*;
+
+import static duke.list.*;
+
+public class Main {
+    public static void main(String[] args){
+        greet();
+        startList();
+        goodbye();
+    }
+}
