+++ src/main/java/duke/Event.java
@@ -1,21 +1,29 @@
+package duke;
+
 import java.util.ArrayList;
 import java.util.List;
 
-public class ToDo extends Task{
+public class Event extends Task{
 
-    public ToDo(String description) throws MissingParamsException {
+    /** Start date of duke.Event */
+    protected String start;
+    /** End date of duke.Event */
+    protected String end;
+    
+    public Event(String description, String start, String end) throws MissingParamsException{
         super(description);
+        this.start = start;
+        this.end = end;
 
-        if (hasMissingParams()) {
+        if(hasMissingParams()) {
             List<TaskParams> missingParams = findMissingParams();
             throw new MissingParamsException(missingParams);
         }
     }
-
     @Override
     public String toString() {
-        return String.format("[T][%s] %s",
-                this.getStatusIcon(), this.description);
+        return String.format("[E][%s] %s (from: %s to: %s)",
+                this.getStatusIcon(), this.description, this.start, this.end);
     }
 
     /**
@@ -30,6 +38,12 @@ public class ToDo extends Task{
 
         if (description.isEmpty()) {
             missingParams.add(TaskParams.DESCRIPTION);
+        }
+        if (start.isEmpty()) {
+            missingParams.add(TaskParams.START);
+        }
+        if (end.isEmpty()) {
+            missingParams.add(TaskParams.END);
         }
 
         return missingParams;
