+++ src/main/java/duke/MissingParamsException.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.List;
 
 public class MissingParamsException extends Exception {
