+++ src/main/java/duke/Task.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.List;
 
 public abstract class Task {
