+++ src/main/java/duke/TaskParams.java
@@ -1,3 +1,5 @@
+package duke;
+
 public enum TaskParams {
     DESCRIPTION, BY, START, END
 }
