+++ src/main/java/duke/ToDo.java
@@ -1,27 +1,23 @@
+package duke;
+
 import java.util.ArrayList;
 import java.util.List;
 
-public class Event extends Task{
+public class ToDo extends Task{
 
-    /** Start date of Event */
-    protected String start;
-    /** End date of Event */
-    protected String end;
-    
-    public Event(String description, String start, String end) throws MissingParamsException{
+    public ToDo(String description) throws MissingParamsException {
         super(description);
-        this.start = start;
-        this.end = end;
 
-        if(hasMissingParams()) {
+        if (hasMissingParams()) {
             List<TaskParams> missingParams = findMissingParams();
             throw new MissingParamsException(missingParams);
         }
     }
+
     @Override
     public String toString() {
-        return String.format("[E][%s] %s (from: %s to: %s)",
-                this.getStatusIcon(), this.description, this.start, this.end);
+        return String.format("[T][%s] %s",
+                this.getStatusIcon(), this.description);
     }
 
     /**
@@ -36,12 +32,6 @@ public class Event extends Task{
 
         if (description.isEmpty()) {
             missingParams.add(TaskParams.DESCRIPTION);
-        }
-        if (start.isEmpty()) {
-            missingParams.add(TaskParams.START);
-        }
-        if (end.isEmpty()) {
-            missingParams.add(TaskParams.END);
         }
 
         return missingParams;
